refactor : #3 충돌에 따른 변경
