+++ src/main/java/hamkke/board/model/user/vo/Password.java
@@ -13,7 +13,7 @@ import java.util.regex.Pattern;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class Password {
 
-    private static final String PASSWORD_REGEX = "^([a-z0-9!@#$%^&*_-]){8,18}$";
+    private static final String PASSWORD_REGEX = "^(?=.*[A-Za-z])(?=.*\\d)(?=.*[!@#$%^&*_-])[A-Za-z\\d!@#$%^&*_-]{8,18}$";
     private static final Pattern PASSWORD_PATTERN = Pattern.compile(PASSWORD_REGEX);
 
     @Column(name = "password")
